wrote DEQs2
Added in a better measure of error of update step to track convergence.
Also removing data points that show a high error and replacing them with
NaNs for plotting here. Still unclear how well this does at getting the
correct cavity prediction - there may be some spread and we may not
always have convergence so may need to check the RS equations to see if
there are multiple finite solutions.